moved process_print_job to printer_utils
